refactor: create utils module for reusable funcs
