+++ tests/test_updaters.py
@@ -1653,104 +1653,6 @@ class TestAttachments:
 
 class TestFeatureServiceOverwriter:
 
-    def test_rename_columns_from_agol_handles_special_and_space(self):
-        cols = ['Test Name:']
-
-        renamed = palletjack.FeatureServiceOverwriter._rename_columns_for_agol(cols)
-
-        assert renamed == {'Test Name:': 'Test_Name_'}
-
-    def test_rename_columns_from_agol_handles_special_and_space_leaves_others_alone(self):
-        cols = ['Test Name:', 'FooName']
-
-        renamed = palletjack.FeatureServiceOverwriter._rename_columns_for_agol(cols)
-
-        assert renamed == {'Test Name:': 'Test_Name_', 'FooName': 'FooName'}
-
-    def test_rename_columns_from_agol_retains_underscores(self):
-        cols = ['Test Name:', 'Foo_Name']
-
-        renamed = palletjack.FeatureServiceOverwriter._rename_columns_for_agol(cols)
-
-        assert renamed == {'Test Name:': 'Test_Name_', 'Foo_Name': 'Foo_Name'}
-
-    def test_check_fields_match_normal(self, mocker):
-        mock_fl = mocker.Mock()
-        mock_fl.properties = {
-            'fields': [
-                {
-                    'name': 'Foo'
-                },
-                {
-                    'name': 'Bar'
-                },
-            ]
-        }
-        df = pd.DataFrame(columns=['Foo', 'Bar'])
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        palletjack.FeatureServiceOverwriter._check_fields_match(overwriter, mock_fl, df)
-
-    def test_check_fields_match_raises_error_on_extra_new_field(self, mocker):
-        mock_fl = mocker.Mock()
-        mock_fl.properties = {
-            'fields': [
-                {
-                    'name': 'Foo'
-                },
-                {
-                    'name': 'Bar'
-                },
-            ]
-        }
-        df = pd.DataFrame(columns=['Foo', 'Bar', 'Baz'])
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        with pytest.raises(RuntimeError) as exc_info:
-            palletjack.FeatureServiceOverwriter._check_fields_match(overwriter, mock_fl, df)
-
-        assert exc_info.value.args[
-            0] == 'New dataset contains the following fields that are not present in the live dataset: {\'Baz\'}'
-
-    def test_check_fields_match_ignores_new_shape_field(self, mocker):
-        mock_fl = mocker.Mock()
-        mock_fl.properties = {
-            'fields': [
-                {
-                    'name': 'Foo'
-                },
-                {
-                    'name': 'Bar'
-                },
-            ]
-        }
-        df = pd.DataFrame(columns=['Foo', 'Bar', 'SHAPE'])
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        palletjack.FeatureServiceOverwriter._check_fields_match(overwriter, mock_fl, df)
-
-    def test_check_fields_match_warns_on_missing_new_field(self, mocker, caplog):
-        mock_fl = mocker.Mock()
-        mock_fl.properties = {
-            'fields': [
-                {
-                    'name': 'Foo'
-                },
-                {
-                    'name': 'Bar'
-                },
-                {
-                    'name': 'Baz'
-                },
-            ]
-        }
-        df = pd.DataFrame(columns=['Foo', 'Bar'])
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        palletjack.FeatureServiceOverwriter._check_fields_match(overwriter, mock_fl, df)
-
-        assert 'New dataset does not contain the following fields that are present in the live dataset: {\'Baz\'}' in caplog.text
-
     def test_truncate_existing_data_normal(self, mocker):
         fl_mock = mocker.Mock()
         fl_mock.manager.truncate.return_value = {
@@ -1788,86 +1690,6 @@ class TestFeatureServiceOverwriter:
         overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
 
         palletjack.FeatureServiceOverwriter._truncate_existing_data(overwriter, fl_mock, 0, 'abc')
-
-    def test_retry_returns_on_first_success(self, mocker):
-        mock = mocker.Mock()
-        mock.function.return_value = 42
-
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        answer = overwriter._retry(mock.function, 'a', 'b')
-
-        assert answer == 42
-        assert mock.function.call_count == 1
-
-    def test_retry_returns_after_one_failure(self, mocker):
-        mock = mocker.Mock()
-        mock.function.side_effect = [urllib.error.HTTPError('a', 'b', 'c', 'd', 'e'), 42]
-        mocker.patch('palletjack.updaters.sleep')
-
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        answer = overwriter._retry(mock.function, 'a', 'b')
-
-        assert answer == 42
-        assert mock.function.call_count == 2
-
-    def test_retry_returns_after_two_failures(self, mocker):
-        mock = mocker.Mock()
-        mock.function.side_effect = [
-            urllib.error.HTTPError('a', 'b', 'c', 'd', 'e'),
-            urllib.error.HTTPError('a', 'b', 'c', 'd', 'e'),
-            42,
-        ]
-        mocker.patch('palletjack.updaters.sleep')
-
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        answer = overwriter._retry(mock.function, 'a', 'b')
-
-        assert answer == 42
-        assert mock.function.call_count == 3
-
-    def test_retry_fails_after_four_failures(self, mocker):
-        mock = mocker.Mock()
-        mock.function.side_effect = [
-            urllib.error.HTTPError('a', 'b', 'c', 'd', 'e'),
-            urllib.error.HTTPError('a', 'b', 'c', 'd', 'e'),
-            urllib.error.HTTPError('a', 'b', 'c', 'd', 'e'),
-            urllib.error.HTTPError('a', 'b', 'c', 'd', 'e'),
-            42,
-        ]
-        mocker.patch('palletjack.updaters.sleep')
-
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        with pytest.raises(urllib.error.HTTPError):
-            answer = overwriter._retry(mock.function, 'a', 'b')
-
-        # assert answer == 42
-        assert mock.function.call_count == 4
-
-    def test_retry_retries_on_request_http_error(self, mocker):
-        mock = mocker.Mock()
-        mock.function.side_effect = [requests.exceptions.HTTPError, 42]
-        mocker.patch('palletjack.updaters.sleep')
-
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        answer = overwriter._retry(mock.function, 'a', 'b')
-
-        assert answer == 42
-
-    def test_retry_retries_on_request_ssl_error(self, mocker):
-        mock = mocker.Mock()
-        mock.function.side_effect = [requests.exceptions.SSLError, 42]
-        mocker.patch('palletjack.updaters.sleep')
-
-        overwriter = palletjack.FeatureServiceOverwriter(mocker.Mock())
-
-        answer = overwriter._retry(mock.function, 'a', 'b')
-
-        assert answer == 42
 
     def test_append_new_data_doesnt_raise_on_normal(self, mocker):
         mock_df = mocker.Mock()
@@ -2006,51 +1828,6 @@ class TestFeatureServiceOverwriter:
             0] == 'Failed to re-add truncated data after failed append; data saved to /foo/bar.json'
         assert 'Append failed; attempting to re-load truncated data...' in caplog.text
         assert 'features reloaded' not in caplog.text
-
-    def test_replace_nan_series_with_empty_strings_one_empty_one_non_empty_float(self, mocker):
-        df = pd.DataFrame({
-            'normal': [1., 2., 3.],
-            'empty': [np.nan, np.nan, np.nan],
-        })
-
-        fixed_df = palletjack.FeatureServiceOverwriter._replace_nan_series_with_empty_strings(mocker.Mock(), df)
-
-        test_df = pd.DataFrame({
-            'normal': [1., 2., 3.],
-            'empty': ['', '', ''],
-        })
-
-        tm.assert_frame_equal(fixed_df, test_df)
-
-    def test_replace_nan_series_with_empty_strings_other_series_has_nan(self, mocker):
-        df = pd.DataFrame({
-            'normal': [1., 2., np.nan],
-            'empty': [np.nan, np.nan, np.nan],
-        })
-
-        fixed_df = palletjack.FeatureServiceOverwriter._replace_nan_series_with_empty_strings(mocker.Mock(), df)
-
-        test_df = pd.DataFrame({
-            'normal': [1., 2., np.nan],
-            'empty': ['', '', ''],
-        })
-
-        tm.assert_frame_equal(fixed_df, test_df)
-
-    def test_replace_nan_series_with_empty_strings_other_series_is_empty_str(self, mocker):
-        df = pd.DataFrame({
-            'normal': ['', '', ''],
-            'empty': [np.nan, np.nan, np.nan],
-        })
-
-        fixed_df = palletjack.FeatureServiceOverwriter._replace_nan_series_with_empty_strings(mocker.Mock(), df)
-
-        test_df = pd.DataFrame({
-            'normal': ['', '', ''],
-            'empty': ['', '', ''],
-        })
-
-        tm.assert_frame_equal(fixed_df, test_df)
 
     def test_save_truncated_data_calls_write_with_json(self, mocker):
         mock_df = pd.DataFrame({
